refactor: html, css, js의 전통적 구조로 파일 분리
