+++ styles/components.css
@@ -1,0 +1,90 @@
+@import url("./colors.css");
+
+.popular-movies-layout {
+  margin: 64px 0;
+  max-width: 1280px;
+  padding: 0 24px;
+  width: 100%;
+}
+
+#popular-movies-title {
+  font-size: 1.4rem;
+  font-weight: bold;
+  margin-bottom: 32px;
+}
+
+.movie-card-layout {
+  display: grid;
+  gap: 48px 16px;
+  grid-template-columns: repeat(5, 1fr); /* 기본: 데스크탑 5개 */
+  padding-bottom: 48px;
+}
+
+@media (width <= 1120px) {
+  .movie-card-layout {
+    grid-template-columns: repeat(4, 1fr); /* 데스크탑 화면 축소 4개 */
+  }
+}
+
+@media (width <= 800px) {
+  .movie-card-layout {
+    grid-template-columns: repeat(3, 1fr); /* 테블릿 3개 */
+  }
+}
+
+@media (width <= 390px) {
+  .movie-card-layout {
+    grid-template-columns: 1fr; /* 모바일 1개 */
+  }
+}
+
+.movie-card {
+  display: flex;
+  flex-direction: column;
+}
+
+.movie-card > .poster {
+  aspect-ratio: 2 / 3;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  border-radius: 8px;
+  margin-bottom: 8px;
+  width: 100%;
+}
+
+.movie-card > .rate {
+  color: var(--color-yellow);
+  top: -1px;
+}
+
+.movie-card > .rate::before {
+  background-image: url("../images/star_empty.png");
+  background-repeat: no-repeat;
+  background-size: contain;
+  content: "";
+  display: inline-block;
+  height: 16px;
+  margin-right: 5px;
+  position: relative;
+  top: -2px;
+  vertical-align: middle;
+  width: 16px;
+}
+
+.movie-card > .title {
+  color: var(--color-white);
+  font-weight: bold;
+}
+
+#load-more-movies {
+  background-color: var(--color-lightblue-90);
+  border: none;
+  border-radius: 8px;
+  color: var(--color-white);
+  cursor: pointer;
+  font-size: 1.2rem;
+  font-weight: bold;
+  height: 46px;
+  width: 100%;
+}
