+++ styles/index.css
@@ -1,6 +1,7 @@
 @import url("./reset.css");
 @import url("./colors.css");
 @import url("./main.css");
+@import url("./components.css");
 @import url("./modal.css");
 @import url("./tab.css");
 @import url("./thumbnail.css");
