+++ templates/index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="./styles/main.css" />
     <link rel="stylesheet" href="./styles/tab.css" />
     <link rel="stylesheet" href="./styles/thumbnail.css" />
+    <link rel="stylesheet" href="./styles/modal.css" />
     <title>영화 리뷰</title>
   </head>
   <body>
